CSP-1479 Implements Update Providers Data function

+++ src/SFA.DAS.PR.Jobs/Models/RegisteredProviderInfo.cs
@@ -1,3 +1,3 @@
 ﻿namespace SFA.DAS.PR.Jobs.Models;
 
-public record RegisteredProviderInfo(int Ukprn, string LegalName);
+public record RegisteredProviderInfo(long Ukprn, string LegalName);
